pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "container-craft"
version = "0.1.0"
description = "A modular Minecraft server orchestration tool using Docker"
authors = [
    { name="m_jimmer", email="josephjamesmills@gmail.com" }
]
readme = "README.md"
license = { file = "LICENCE.md" }
dependencies = [
    "pyyaml",
    "jsonschema",
    "gitpython",
    "httpx",
    "docker",
    "kconfiglib",
    "jinja2"
]
requires-python = ">=3.8"

[project.scripts]
container-craft = "container_craft:main"
mcpkg = "mcpkg:main"

[tool.pytest.ini_options]
pythonpath = ["."]
